hypervector/requirements.txt
ploomber
pandas
scikit-learn
sklearn-evaluation
